+++ tools/link.py
@@ -2721,7 +2721,6 @@ def map_to_js_libs(library_name):
     'dl': [],
     'm': [],
     'rt': [],
-    'pthread': [],
     # This is the name of GNU's C++ standard library. We ignore it here
     # for compatibility with GNU toolchains.
     'stdc++': [],